Replace size_t with uint64_t in API methods

+++ bindings/nim/kzg_abi.nim
@@ -91,16 +91,16 @@ type
 
 proc load_trusted_setup*(res: KzgSettings,
                          g1MonomialBytes: ptr byte,
-                         numG1MonomialBytes: csize_t,
+                         numG1MonomialBytes: uint64,
                          g1LagrangeBytes: ptr byte,
-                         numG1LagrangeBytes: csize_t,
+                         numG1LagrangeBytes: uint64,
                          g2MonomialBytes: ptr byte,
-                         numG2MonomialBytes: csize_t,
-                         precompute: csize_t): KZG_RET {.kzg_abi.}
+                         numG2MonomialBytes: uint64,
+                         precompute: uint64): KZG_RET {.kzg_abi.}
 
 proc load_trusted_setup_file*(res: KzgSettings,
                          input: File,
-                         precompute: csize_t): KZG_RET {.kzg_abi.}
+                         precompute: uint64): KZG_RET {.kzg_abi.}
 
 proc free_trusted_setup*(s: KzgSettings) {.kzg_abi.}
 
@@ -136,7 +136,7 @@ proc verify_blob_kzg_proof_batch*(res: var bool,
                          blobs: ptr KzgBlob,
                          commitmentsBytes: ptr KzgBytes48,
                          proofBytes: ptr KzgBytes48,
-                         n: csize_t,
+                         n: uint64,
                          s: KzgSettings): KZG_RET {.kzg_abi.}
 
 proc compute_cells_and_kzg_proofs*(cellsOut: ptr KzgCell,
@@ -148,7 +148,7 @@ proc recover_cells_and_kzg_proofs*(recoveredOut: ptr KzgCell,
                          recoveredProofsOut: ptr KzgProof,
                          cellIndices: ptr uint64,
                          cells: ptr KzgCell,
-                         numCells: csize_t,
+                         numCells: uint64,
                          s: KzgSettings): KZG_RET {.kzg_abi.}
 
 proc verify_cell_kzg_proof_batch*(res: var bool,
@@ -156,5 +156,5 @@ proc verify_cell_kzg_proof_batch*(res: var bool,
                          cellIndices: ptr uint64,
                          cells: ptr KzgCell,
                          proofs: ptr KzgBytes48,
-                         numCells: csize_t,
+                         numCells: uint64,
                          s: KzgSettings): KZG_RET {.kzg_abi.}
